pid is now long

+++ src/main/java/de/flapdoodle/embed/memcached/MemcachedProcess.java
@@ -149,7 +149,7 @@ public class MemcachedProcess extends
 			// force to find out the pid anyways. This works pretty well
 			// on Unix
 			// (won't help us much here..) and needs Sigar lib on Windows.
-			setProcessId(process.getPid().intValue());
+			setProcessId(process.getPid());
 		} else {
 			setProcessId(getPidFromFile(pidFile()));
 		}
